+++ inbm-vision/packaging/build-agent-exe-py3.sh
@@ -62,7 +62,7 @@ fpm -t $2\
     -n ${NAME}\
     -v ${VERSION}\
     --no-auto-depends\
-    --depends lxc-common\
+    --depends lxc\
     -a all\
     -m none\
     .