+++ inbm/Changelog.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
  - RTC 517782 - Add Ubuntu 22.04 support for INBM TC including integration test
+ - RTC 517781 - Allow installation on Debian bullseye
 
 ## 3.0.12 - 2022-07-14